Working on the front end Moving logic from App.js into a Component that will render the loading page. Moved the initial input to be handled by a functional component called Input. Running into a CORS issue that will need to be addressed. Moved API call into its own file
